Add files via upload
Verily, I had crafted a new footer, though it remained not entirely complete. Nevertheless, I am compelled to commit it, owing to the pressing matter of a milestone. The previous footer, , was not to my liking nor did it please my discerning eye. Thus, I embarked upon this endeavor to create a more fitting conclusion to my work.